Move sim-specific container stuff to separate module
