+++ crates/sui-swarm/src/memory/node.rs
@@ -3,17 +3,12 @@
 
 use anyhow::anyhow;
 use anyhow::Result;
-use futures::FutureExt;
-use prometheus::Registry;
-use std::thread;
 use sui_config::NodeConfig;
-use sui_node::SuiNode;
 use sui_types::base_types::SuiAddress;
 use tap::TapFallible;
 use tracing::{error, trace};
 
-#[cfg(msim)]
-use std::net::{IpAddr, SocketAddr};
+use super::container::Container;
 
 /// A handle to an in-memory Sui Node.
 ///
@@ -114,185 +109,6 @@ impl std::fmt::Display for HealthCheckError {
 }
 
 impl std::error::Error for HealthCheckError {}
-
-#[derive(Debug)]
-struct Container {
-    join_handle: Option<ContainerJoinHandle>,
-    cancel_sender: Option<tokio::sync::oneshot::Sender<()>>,
-}
-
-#[derive(Debug)]
-enum ContainerJoinHandle {
-    #[allow(dead_code)]
-    Thread(thread::JoinHandle<()>),
-
-    #[cfg(msim)]
-    SimulatorNode(
-        sui_simulator::task::NodeId,
-        sui_simulator::task::JoinHandle<()>,
-    ),
-}
-
-/// When dropped, stop and wait for the node running in this Container to completely shutdown.
-impl Drop for Container {
-    fn drop(&mut self) {
-        trace!("dropping Container");
-
-        let join_handle = self.join_handle.take().unwrap();
-
-        match join_handle {
-            ContainerJoinHandle::Thread(thread) => {
-                let cancel_handle = self.cancel_sender.take().unwrap();
-
-                // Notify the thread to shutdown
-                let _ = cancel_handle.send(());
-
-                // Wait for the thread to join
-                thread.join().unwrap();
-            }
-
-            #[cfg(msim)]
-            ContainerJoinHandle::SimulatorNode(node_id, join_handle) => {
-                tracing::info!("shutting down {}", node_id);
-                join_handle.abort();
-                sui_simulator::runtime::Handle::try_current().map(|h| h.kill(node_id));
-            }
-        }
-
-        trace!("finished dropping Container");
-    }
-}
-
-impl Container {
-    /// Spawn a new Node.
-    pub fn spawn(
-        config: NodeConfig,
-        _runtime: RuntimeType,
-    ) -> (tokio::sync::oneshot::Receiver<()>, Self) {
-        #[cfg(msim)]
-        return Self::spawn_simulator_nodes(config);
-
-        #[cfg(not(msim))]
-        return Self::spawn_threads(config, _runtime);
-    }
-
-    #[cfg(msim)]
-    fn spawn_simulator_nodes(config: NodeConfig) -> (tokio::sync::oneshot::Receiver<()>, Self) {
-        let (startup_sender, startup_reciever) = tokio::sync::oneshot::channel();
-        let (cancel_sender, cancel_reciever) = tokio::sync::oneshot::channel();
-
-        let handle = sui_simulator::runtime::Handle::current();
-        let builder = handle.create_node();
-
-        let socket_addr =
-            mysten_network::multiaddr::to_socket_addr(&config.network_address).unwrap();
-        let ip = match socket_addr {
-            SocketAddr::V4(v4) => IpAddr::V4(*v4.ip()),
-            _ => panic!("unsupported protocol"),
-        };
-
-        let node = builder
-            .ip(ip)
-            .name(format!("{}", config.protocol_public_key()))
-            .init(|| async {
-                tracing::info!("node restarted");
-            })
-            .build();
-
-        let join_handle = node.spawn(async move {
-            let _server = SuiNode::start(&config).await.unwrap();
-            // Notify that we've successfully started the node
-            error!("node started, sending oneshot");
-            let _ = startup_sender.send(());
-            // run until canceled
-            cancel_reciever.map(|_| ()).await;
-            trace!("cancellation received; shutting down thread");
-        });
-
-        (
-            startup_reciever,
-            Self {
-                join_handle: Some(ContainerJoinHandle::SimulatorNode(node.id(), join_handle)),
-                cancel_sender: Some(cancel_sender),
-            },
-        )
-    }
-
-    #[cfg(not(msim))]
-    fn spawn_threads(
-        config: NodeConfig,
-        runtime: RuntimeType,
-    ) -> (tokio::sync::oneshot::Receiver<()>, Self) {
-        let (startup_sender, startup_reciever) = tokio::sync::oneshot::channel();
-        let (cancel_sender, cancel_reciever) = tokio::sync::oneshot::channel();
-
-        let thread = thread::spawn(move || {
-            let span = tracing::span!(
-                tracing::Level::INFO,
-                "node",
-                name =% config.sui_address()
-            );
-            let _guard = span.enter();
-
-            let mut builder = match runtime {
-                RuntimeType::SingleThreaded => tokio::runtime::Builder::new_current_thread(),
-                RuntimeType::MultiThreaded => {
-                    thread_local! {
-                        static SPAN: std::cell::RefCell<Option<tracing::span::EnteredSpan>> =
-                            std::cell::RefCell::new(None);
-                    }
-                    let mut builder = tokio::runtime::Builder::new_multi_thread();
-                    let span = span.clone();
-                    builder
-                        .on_thread_start(move || {
-                            SPAN.with(|maybe_entered_span| {
-                                *maybe_entered_span.borrow_mut() = Some(span.clone().entered());
-                            });
-                        })
-                        .on_thread_stop(|| {
-                            SPAN.with(|maybe_entered_span| {
-                                maybe_entered_span.borrow_mut().take();
-                            });
-                        });
-
-                    builder
-                }
-            };
-            let runtime = builder.enable_all().build().unwrap();
-            runtime.block_on(async move {
-                let _server = SuiNode::start(&config, Registry::new()).await.unwrap();
-                // Notify that we've successfully started the node
-                let _ = startup_sender.send(());
-                // run until canceled
-                cancel_reciever.map(|_| ()).await;
-
-                trace!("cancellation received; shutting down thread");
-            });
-        });
-
-        (
-            startup_reciever,
-            Self {
-                join_handle: Some(ContainerJoinHandle::Thread(thread)),
-                cancel_sender: Some(cancel_sender),
-            },
-        )
-    }
-
-    /// Check to see that the Node is still alive by checking if the receiving side of the
-    /// `cancel_sender` has been dropped.
-    ///
-    //TODO When we move to rust 1.61 we should also use
-    // https://doc.rust-lang.org/stable/std/thread/struct.JoinHandle.html#method.is_finished
-    // in order to check if the thread has finished.
-    pub fn is_alive(&self) -> bool {
-        if let Some(cancel_sender) = &self.cancel_sender {
-            !cancel_sender.is_closed()
-        } else {
-            false
-        }
-    }
-}
 
 /// The type of tokio runtime that should be used for a particular Node
 #[derive(Clone, Copy, Debug)]
